automatically backed up by learn

+++ db/migrate/20140805183012_create_items.rb
@@ -1,12 +1,13 @@
 class CreateItems < ActiveRecord::Migration
   def change
     create_table :items do |t|
-      		t.string :title
-      		t.integer :inventory
-      		t.integer :price
-      		t.integer :category_id
-
-      		t.timestamps null: false
-      	end
+       t.string :title
+       t.integer :inventory
+       t.float :price
+       t.integer :category_id
+ 
+       t.timestamps null: false
+      		
+    end
   end
 end
